+++ resolwe_bio/tools/bamtobigwig.sh
@@ -2,13 +2,18 @@
 
 source /etc/profile.d/resolwe-base.sh
 
-NAME=`basename "$1" .bam`
+bam_file="$1"
+number_of_cores="$2"
+bin_size="$3"
 
-samtools idxstats "$1" | cut -f -2 | head -n -1 > chrom.sizes
-re-checkrc
-genomeCoverageBed -bg -ibam "$1" > tmp.bedgraph
-re-checkrc "Bam to Bedgraph failed."
-bedGraphToBigWig tmp.bedgraph chrom.sizes "${NAME}.bw"
-re-checkrc "BedGraph to BigWig failed."
+NAME=`basename "${bam_file}" .bam`
+
+bamCoverage \
+  --binSize "${bin_size}" \
+  --bam "${bam_file}" \
+  --outFileName "${NAME}.bw" \
+  --outFileFormat "bigwig" \
+  --numberOfProcessors "${number_of_cores}"
+re-checkrc "Bam to BigWig failed."
 
 re-save-file bigwig "${NAME}.bw"